pyproject.toml
[tool.poetry]
name = "dashboard"
version = "0.1.0"
description = "Dashboard"
authors = ["You"]
readme = "README.md"

packages = [
    { include = "auth" },
    { include = "VPN" },
    { include = "Codex" },
    { include = "database" },
    { include = "config" }
]

[tool.poetry.dependencies]
python = "^3.12"
asyncpg = "^0.29.0"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
sqlalchemy = "^2.0"
python-dotenv = "^1.0.1"
jinja2 = "^3.1.3"
itsdangerous = "^2.2.0"
python-multipart = "^0.0.9"
psycopg2 = "^2.9.11"

[tool.poetry.group.dev.dependencies]
black = "^24.8"
mypy = "^1.10"
pre-commit = "^3.8"
sqlalchemy2-stubs = "0.0.2a38"
ruff = "^0.8.0"

[tool.black]
line-length = 79
target-version = ['py310']
skip-string-normalization = true

[tool.ruff]
line-length = 79
target-version = "py310"

exclude = ["venv", ".venv", "build", "dist", "__pycache__"]

[tool.ruff.lint]
select = ["E", "W", "F", "I"]

ignore = ["E203"]


fixable = ["F401", "I"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
strict = false
warn_unused_ignores = true
warn_return_any = false
warn_unreachable = false
disable_error_code = ["attr-defined", "misc", "call-arg", "assignment"]

[[tool.mypy.overrides]]
module = "sqlalchemy.*"
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
